print list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -11,7 +11,7 @@
 
 size_t print_list(const list_t *h)
 {
-	size_t size = 0;
+	unsigned int size = 0;
 
 	while (h)
 	{
